Split tests and spec_helper into separate files
