+++ spec/support/node.rb
@@ -1,0 +1,24 @@
+require 'arboreal'
+
+class Node < ActiveRecord::Base
+  acts_arboreal
+
+  class Migration < ActiveRecord::Migration
+    def self.up
+      create_table "nodes", :force => true do |t|
+        t.string "name"
+        t.string "type"
+        t.integer "parent_id"
+        t.string "ancestry_string"
+      end
+    end
+
+    def self.down
+      drop_table "nodes"
+    end
+  end
+end
+
+class RedNode < Node; end
+class GreenNode < Node; end
+class BlueNode < Node; end
